move section checks to each individual section test
