+++ internal/pkg/message/message_test.go
@@ -19,7 +19,11 @@ func CheckMessage(m1, m2 Message, t *testing.T) {
 			t.Error("Capabilities mismatch")
 		}
 	}
-	section.CheckSignatures(m1.Signatures, m2.Signatures, t)
+	for i, s1 := range m1.Signatures {
+		if s1 != m2.Signatures[i] {
+			t.Fatalf("Assertions are not equal q1=%s q2=%s", s1, m2.Signatures[i])
+		}
+	}
 	if len(m1.Content) != len(m2.Content) {
 		t.Error("Message Content length mismatch")
 	}
@@ -27,33 +31,41 @@ func CheckMessage(m1, m2 Message, t *testing.T) {
 		switch s1 := s1.(type) {
 		case *section.Assertion:
 			if s2, ok := m2.Content[i].(*section.Assertion); ok {
-				section.CheckAssertion(s1, s2, t)
+				if s1.CompareTo(s2) != 0 {
+					t.Fatalf("Assertions are not equal q1=%s q2=%s", s1, s2)
+				}
 				continue
 			}
 			t.Errorf("Types at position %d of Content slice are different", i)
 		case *section.Shard:
 			if s2, ok := m2.Content[i].(*section.Shard); ok {
-				section.CheckShard(s1, s2, t)
+				if s1.CompareTo(s2) != 0 {
+					t.Fatalf("Shards are not equal q1=%s q2=%s", s1, s2)
+				}
 				continue
 			}
 			t.Errorf("Types at position %d of Content slice are different", i)
 		case *section.Zone:
 			if s2, ok := m2.Content[i].(*section.Zone); ok {
-				section.CheckZone(s1, s2, t)
+				if s1.CompareTo(s2) != 0 {
+					t.Fatalf("Zones are not equal q1=%s q2=%s", s1, s2)
+				}
 				continue
 			}
 			t.Errorf("Types at position %d of Content slice are different", i)
 		case *query.Name:
 			if s2, ok := m2.Content[i].(*query.Name); ok {
 				if s1.CompareTo(s2) != 0 {
-					t.Errorf("Queries are not equal q1=%s q2=%s", s1, s2)
+					t.Fatalf("Queries are not equal q1=%s q2=%s", s1, s2)
 				}
 				continue
 			}
 			t.Errorf("Types at position %d of Content slice are different", i)
 		case *section.Notification:
 			if s2, ok := m2.Content[i].(*section.Notification); ok {
-				section.CheckNotification(s1, s2, t)
+				if s1.CompareTo(s2) != 0 {
+					t.Fatalf("Notifications are not equal q1=%s q2=%s", s1, s2)
+				}
 				continue
 			}
 			t.Errorf("Types at position %d of Content slice are different", i)
